+++ Sources/VComponents/Components/Modals/Bottom Sheet/VBottomSheet.swift
@@ -106,6 +106,7 @@ struct VBottomSheet<HeaderLabel, Content>: View
                     divider
                 })
                     .readSize(onChange: { headerDividerHeight = $0.height })
+                    .safeAreaMarginInsets(edges: uiModel.layout.headerSafeAreaEdges)
 
                 contentView
             })